requirements.txt
celery==4.4.2
Django==3.0.4
django-debug-toolbar==2.2
django-extensions==2.2.8
django-rest-swagger==2.2.0
django-tables2==2.2.1
django-widget-tweaks==1.4.8
django-formtools==2.2
django-bootstrap-form==3.4
djangorestframework==3.11.0
mysqlclient==1.4.6
flower==0.9.3
redis==3.4.1
requests==2.23.0
tornado==6.0.4
lxml==4.5.0
Pairtree==0.8.1
pika==1.1.0
xmltodict==0.12.0
py-dateutil==2.2
pycountry==19.8.18
python-magic==0.4.15
rdflib==4.2.2
drf-yasg==1.17.1
djangorestframework-api-key==1.4.1
simplejson
-e git://github.com/E-ARK-Software/eatb.git#egg=eatb
#-e ../eatb